update dive-deco to 2.0.0

+++ src/common.rs
@@ -1,4 +1,4 @@
 pub type Depth = f64;
-pub type Seconds = usize;
+pub type Seconds = u64;
 pub type GF = f64;
 pub type GradientFactorsSetting = (u8, u8);
